# api/swagger.yaml
basePath: /beetle
definitions:
...
  controller.MigrateInfraRequest:
    properties:
      description:
        type: string
      status:
        type: string
      targetSecurityGroupList:
        items:
          $ref: '#/definitions/model.TbSecurityGroupReq'
        type: array
      targetVNet:
        $ref: '#/definitions/model.TbVNetReq'
      targetVmInfra:
        $ref: '#/definitions/model.TbMciReq'
      targetVmOsImageList:
        items:
          $ref: '#/definitions/model.TbImageInfo'
        type: array
      targetVmSpecList:
        items:
          $ref: '#/definitions/model.TbSpecInfo'
        type: array
    type: object
  controller.MigrateInfraResponse:
    type: object
  controller.MigrateInfraWithDefaultsRequest:
    properties:
      description:
        example: Made in CB-TB
        type: string
      installMonAgent:
        default: "no"
        description: InstallMonAgent Option for CB-Dragonfly agent installation ([yes/no]
          default:no)
        enum:
        - "yes"
        - "no"
        example: "no"
        type: string
      label:
        additionalProperties:
          type: string
        description: Label is for describing the object by keywords
        type: object
      name:
        example: mci01
        type: string
      postCommand:
        allOf:
        - $ref: '#/definitions/model.MciCmdReq'
        description: PostCommand is for the command to bootstrap the VMs
      systemLabel:
        description: SystemLabel is for describing the mci in a keyword (any string
          can be used) for special System purpose
        example: ""
        type: string
      vm:
        items:
          $ref: '#/definitions/model.TbVmDynamicReq'
        type: array
    required:
    - name
    - vm
    type: object
  controller.MigrateInfraWithDefaultsResponse:
    properties:
      configureCloudAdaptiveNetwork:
        default: "no"
        description: ConfigureCloudAdaptiveNetwork is an option to configure Cloud
          Adaptive Network (CLADNet) ([yes/no] default:yes)
        enum:
        - "yes"
        - "no"
        example: "yes"
        type: string
      description:
        type: string
      id:
        description: Id is unique identifier for the object
        example: aws-ap-southeast-1
        type: string
      installMonAgent:
        default: "no"
        description: InstallMonAgent Option for CB-Dragonfly agent installation ([yes/no]
          default:no)
        enum:
        - "yes"
        - "no"
        example: "no"
        type: string
      label:
        additionalProperties:
          type: string
        description: Label is for describing the object by keywords
        type: object
      name:
        description: Name is human-readable string to represent the object
        example: aws-ap-southeast-1
        type: string
      newVmList:
        description: List of IDs for new VMs. Return IDs if the VMs are newly added.
          This field should be used for return body only.
        items:
          type: string
        type: array
      placementAlgo:
        type: string
      postCommand:
        allOf:
        - $ref: '#/definitions/model.MciCmdReq'
        description: PostCommand is for the command to bootstrap the VMs
      postCommandResult:
        allOf:
        - $ref: '#/definitions/model.MciSshCmdResult'
        description: PostCommandResult is the result of the command for bootstraping
          the VMs
      resourceType:
        description: ResourceType is the type of the resource
        type: string
      status:
        type: string
      statusCount:
        $ref: '#/definitions/model.StatusCountInfo'
      systemLabel:
        description: SystemLabel is for describing the mci in a keyword (any string
          can be used) for special System purpose
        example: Managed by CB-Tumblebug
        type: string
      systemMessage:
        description: Latest system message such as error message
        example: Failed because ...
        type: string
      targetAction:
        type: string
      targetStatus:
        type: string
      uid:
        description: Uid is universally unique identifier for the object, used for
          labelSelector
        example: wef12awefadf1221edcf
        type: string
      vm:
        items:
          $ref: '#/definitions/model.TbVmInfo'
        type: array
    type: object
...
  github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response:
    properties:
      details:
        example: Any details
        type: string
      list:
        items: {}
        type: array
      object:
        additionalProperties: true
        type: object
      success:
        example: true
        type: boolean
      text:
        example: Any text
        type: string
    type: object
...
  migration.MciInfoList:
    properties:
      mci:
        items:
          $ref: '#/definitions/model.MciStatusInfo'
        type: array
    type: object
...
  model.KeyValue:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  model.MciCmdReq:
    properties:
      command:
        example:
        - 'client_ip=$(echo $SSH_CLIENT | awk ''{print $1}''); echo SSH client IP
          is: $client_ip'
        items:
          type: string
        type: array
      userName:
        example: cb-user
        type: string
    required:
    - command
    type: object
  model.MciSshCmdResult:
    properties:
      results:
        items:
          $ref: '#/definitions/model.SshCmdResult'
        type: array
    type: object
  model.MciStatusInfo:
    properties:
      id:
        type: string
      installMonAgent:
        description: InstallMonAgent Option for CB-Dragonfly agent installation ([yes/no]
          default:yes)
        example: '[yes, no]'
        type: string
      label:
        additionalProperties:
          type: string
        description: Label is for describing the object by keywords
        type: object
      masterIp:
        example: 32.201.134.113
        type: string
      masterSSHPort:
        type: string
      masterVmId:
        example: vm-asiaeast1-cb-01
        type: string
      name:
        type: string
      status:
        type: string
      statusCount:
        $ref: '#/definitions/model.StatusCountInfo'
      systemLabel:
        description: SystemLabel is for describing the mci in a keyword (any string
          can be used) for special System purpose
        example: Managed by CB-Tumblebug
        type: string
      targetAction:
        type: string
      targetStatus:
        type: string
      vm:
        items:
          $ref: '#/definitions/model.TbVmStatusInfo'
        type: array
    type: object
...
  model.StatusCountInfo:
    properties:
      countCreating:
        description: CountCreating is for counting Creating
        type: integer
      countFailed:
        description: CountFailed is for counting Failed
        type: integer
      countRebooting:
        description: CountRebooting is for counting Rebooting
        type: integer
      countResuming:
        description: CountResuming is for counting Resuming
        type: integer
      countRunning:
        description: CountRunning is for counting Running
        type: integer
      countSuspended:
        description: CountSuspended is for counting Suspended
        type: integer
      countSuspending:
        description: CountSuspending is for counting Suspending
        type: integer
      countTerminated:
        description: CountTerminated is for counting Terminated
        type: integer
      countTerminating:
        description: CountTerminating is for counting Terminating
        type: integer
      countTotal:
        description: CountTotal is for Total VMs
        type: integer
      countUndefined:
        description: CountUndefined is for counting Undefined
        type: integer
    type: object
...
  model.TbImageInfo:
    properties:
      connectionName:
        type: string
      creationDate:
        type: string
      cspImageName:
        example: csp-06eb41e14121c550a
        type: string
      description:
        type: string
      details:
        items:
          $ref: '#/definitions/model.KeyValue'
        type: array
      fetchedTime:
        type: string
      id:
        example: aws-ap-southeast-1
        type: string
      imageStatus:
        allOf:
        - $ref: '#/definitions/model.ImageStatus'
        description: Available, Deprecated, NA
        example: Available
      infraType:
        description: vm|k8s|kubernetes|container, etc.
        type: string
      isGPUImage:
        default: false
        type: boolean
      isKubernetesImage:
        default: false
        type: boolean
      name:
        example: aws-ap-southeast-1
        type: string
      namespace:
        description: Composite primary key
        example: default
        type: string
      osArchitecture:
        allOf:
        - $ref: '#/definitions/model.OSArchitecture'
        description: arm64, x86_64 etc.
        example: x86_64
      osDiskSizeGB:
        description: 10, 50, 100 etc.
        example: 50
        type: number
      osDiskType:
        description: ebs, HDD, etc.
        example: HDD
        type: string
      osDistribution:
        description: Ubuntu 22.04~, CentOS 8 etc.
        example: Ubuntu 22.04~
        type: string
      osPlatform:
        allOf:
        - $ref: '#/definitions/model.OSPlatform'
        description: Linux/UNIX, Windows, NA
        example: Linux/UNIX
      osType:
        example: ubuntu 22.04
        type: string
      providerName:
        type: string
      regionList:
        description: Array field for supporting multiple regions
        items:
          type: string
        type: array
      systemLabel:
        example: Managed by CB-Tumblebug
        type: string
      uid:
        example: wef12awefadf1221edcf
        type: string
    type: object
...
  model.TbMciReq:
    properties:
      description:
        example: Made in CB-TB
        type: string
      installMonAgent:
        default: "no"
        description: InstallMonAgent Option for CB-Dragonfly agent installation ([yes/no]
          default:yes)
        enum:
        - "yes"
        - "no"
        example: "no"
        type: string
      label:
        additionalProperties:
          type: string
        description: Label is for describing the object by keywords
        type: object
      name:
        example: mci01
        type: string
      placementAlgo:
        type: string
      postCommand:
        allOf:
        - $ref: '#/definitions/model.MciCmdReq'
        description: PostCommand is for the command to bootstrap the VMs
      systemLabel:
        description: SystemLabel is for describing the mci in a keyword (any string
          can be used) for special System purpose
        example: ""
        type: string
      vm:
        items:
          $ref: '#/definitions/model.TbVmReq'
        type: array
    required:
    - name
    - vm
    type: object
...
  model.TbSecurityGroupReq:
    properties:
      connectionName:
        type: string
      cspResourceId:
        description: CspResourceId is required to register object from CSP (option=register)
        type: string
      description:
        type: string
      firewallRules:
        description: validate:"required"`
        items:
          $ref: '#/definitions/model.TbFirewallRuleInfo'
        type: array
      name:
        type: string
      vNetId:
        type: string
    required:
    - connectionName
    - name
    - vNetId
    type: object
  model.TbSpecInfo:
    properties:
      acceleratorCount:
        type: integer
      acceleratorMemoryGB:
        type: number
      acceleratorModel:
        type: string
      acceleratorType:
        type: string
      associatedObjectList:
        items:
          type: string
        type: array
      connectionName:
        type: string
      costPerHour:
        type: number
      cspSpecName:
        description: CspSpecName is name of the spec given by CSP
        example: csp-06eb41e14121c550a
        type: string
      description:
        type: string
      diskSizeGB:
        type: number
      evaluationScore01:
        type: number
      evaluationScore02:
        type: number
      evaluationScore03:
        type: number
      evaluationScore04:
        type: number
      evaluationScore05:
        type: number
      evaluationScore06:
        type: number
      evaluationScore07:
        type: number
      evaluationScore08:
        type: number
      evaluationScore09:
        type: number
      evaluationScore10:
        type: number
      evaluationStatus:
        type: string
      id:
        description: Id is unique identifier for the object
        example: aws-ap-southeast-1
        type: string
      infraType:
        description: InfraType can be one of vm|k8s|kubernetes|container, etc.
        type: string
      isAutoGenerated:
        type: boolean
      maxTotalStorageTiB:
        type: integer
      memoryGiB:
        type: number
      name:
        description: Name is human-readable string to represent the object
        example: aws-ap-southeast-1
        type: string
      namespace:
        example: default
        type: string
      netBwGbps:
        type: integer
      orderInFilteredResult:
        type: integer
      osType:
        type: string
      providerName:
        type: string
      regionName:
        type: string
      rootDiskSize:
        type: string
      rootDiskType:
        type: string
      systemLabel:
        description: SystemLabel is for describing the Resource in a keyword (any
          string can be used) for special System purpose
        example: Managed by CB-Tumblebug
        type: string
      uid:
        description: Uid is universally unique identifier for the object, used for
          labelSelector
        example: wef12awefadf1221edcf
        type: string
      vCPU:
        type: integer
    type: object
...
  model.TbVNetReq:
    properties:
      cidrBlock:
        example: 10.0.0.0/16
        type: string
      connectionName:
        example: aws-ap-northeast-2
        type: string
      description:
        example: vnet00 managed by CB-Tumblebug
        type: string
      name:
        example: vnet00
        type: string
      subnetInfoList:
        items:
          $ref: '#/definitions/model.TbSubnetReq'
        type: array
    required:
    - connectionName
    - name
    type: object
  model.TbVmDynamicReq:
    properties:
      commonImage:
        description: CommonImage is field for id of a image in common namespace
        example: ubuntu18.04
        type: string
      commonSpec:
        description: CommonSpec is field for id of a spec in common namespace
        example: aws+ap-northeast-2+t2.small
        type: string
      connectionName:
        description: |-
          if ConnectionName is given, the VM tries to use associtated credential.
          if not, it will use predefined ConnectionName in Spec objects
        type: string
      description:
        example: Description
        type: string
      label:
        additionalProperties:
          type: string
        description: Label is for describing the object by keywords
        type: object
      name:
        description: VM name or subGroup name if is (not empty) && (> 0). If it is
          a group, actual VM name will be generated with -N postfix.
        example: g1-1
        type: string
      rootDiskSize:
        default: default
        description: '"default", Integer (GB): ["50", ..., "1000"]'
        example: default, 30, 42, ...
        type: string
      rootDiskType:
        default: default
        description: '"", "default", "TYPE1", AWS: ["standard", "gp2", "gp3"], Azure:
          ["PremiumSSD", "StandardSSD", "StandardHDD"], GCP: ["pd-standard", "pd-balanced",
          "pd-ssd", "pd-extreme"], ALIBABA: ["cloud_efficiency", "cloud", "cloud_essd"],
          TENCENT: ["CLOUD_PREMIUM", "CLOUD_SSD"]'
        example: default, TYPE1, ...
        type: string
      subGroupSize:
        default: "1"
        description: if subGroupSize is (not empty) && (> 0), subGroup will be generated.
          VMs will be created accordingly.
        example: "3"
        type: string
      vmUserPassword:
        type: string
    required:
    - commonImage
    - commonSpec
    type: object
...
host: localhost:8056
info:
  contact:
    name: API Support
    url: https://github.com/cloud-barista/cm-beetle/issues/new/choose
  description: CM-Beetle REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Beetle REST API
  version: latest
paths:
...
  /migration/ns/{nsId}/mci:
    get:
      consumes:
      - application/json
      description: Get the migrated multi-cloud infrastructure (MCI)
      operationId: ListInfra
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: status
        description: Option for getting the migrated multi-cloud infrastructure
        enum:
        - status
        - id
        in: query
        name: option
        type: string
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The info list of the migrated multi-cloud infrastructure (MCI)
          schema:
            $ref: '#/definitions/migration.MciInfoList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
      summary: Get the migrated multi-cloud infrastructure (MCI)
      tags:
      - '[Migration] Infrastructure'
    post:
      consumes:
      - application/json
      description: Migrate an infrastructure to the multi-cloud infrastructure (MCI)
        with defaults.
      operationId: MigrateInfra
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Specify the information for the targeted mulci-cloud infrastructure
          (MCI)
        in: body
        name: mciInfo
        required: true
        schema:
          $ref: '#/definitions/controller.MigrateInfraRequest'
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated to the multi-cloud infrastructure
          schema:
            $ref: '#/definitions/controller.MigrateInfraResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
      summary: Migrate an infrastructure to the multi-cloud infrastructure (MCI) with
        defaults
      tags:
      - '[Migration] Infrastructure'
  /migration/ns/{nsId}/mci/{mciId}:
    delete:
      consumes:
      - application/json
      description: Delete the migrated mult-cloud infrastructure (MCI)
      operationId: DeleteInfra
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: mmci01
        description: Migrated Multi-Cloud Infrastructure (MCI) ID
        in: path
        name: mciId
        required: true
        type: string
      - default: terminate
        description: Action for deletion
        enum:
        - terminate
        - force
        in: query
        name: action
        type: string
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The result of deleting the migrated multi-cloud infrastructure
            (MCI)
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
      summary: Delete the migrated mult-cloud infrastructure (MCI)
      tags:
      - '[Migration] Infrastructure'
    get:
      consumes:
      - application/json
      description: Get the migrated multi-cloud infrastructure (MCI)
      operationId: GetInfra
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: mmci01
        description: Migrated Multi-Cloud Infrastructure (MCI) ID
        in: path
        name: mciId
        required: true
        type: string
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The migrated multi-cloud infrastructure (MCI) information
          schema:
            $ref: '#/definitions/controller.MigrateInfraResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
      summary: Get the migrated multi-cloud infrastructure (MCI)
      tags:
      - '[Migration] Infrastructure'
  /migration/ns/{nsId}/mciWithDefaults:
    post:
      consumes:
      - application/json
      description: Migrate an infrastructure to the multi-cloud infrastructure (MCI)
        with defaults.
      operationId: MigrateInfraWithDefaults
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Specify the information for the targeted mulci-cloud infrastructure
          (MCI)
        in: body
        name: mciInfo
        required: true
        schema:
          $ref: '#/definitions/controller.MigrateInfraWithDefaultsRequest'
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated to the multi-cloud infrastructure
          schema:
            $ref: '#/definitions/controller.MigrateInfraWithDefaultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
      summary: Migrate an infrastructure to the multi-cloud infrastructure (MCI) with
        defaults
      tags:
      - '[Migration] Infrastructure'
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"